# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Andromeda Educational Program Comparison API",
    "description": "Strict source-backed contracts for comparing BMSTU educational programmes.",
    "version": "1.0.0"
  },
  "paths": {
    "/programs": {
      "get": {
        "tags": [
          "programs"
        ],
        "summary": "List Programs",
        "operationId": "list_programs_programs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/programs/{id}": {
      "get": {
        "tags": [
          "programs"
        ],
        "summary": "Get Program",
        "operationId": "get_program_programs__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^program:[0-9]{2}\\.[0-9]{2}\\.[0-9]{2}-[0-9]{2,3}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/programs/{id}/curriculum": {
      "get": {
        "tags": [
          "programs"
        ],
        "summary": "Get Curriculum",
        "operationId": "get_curriculum_programs__id__curriculum_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^program:[0-9]{2}\\.[0-9]{2}\\.[0-9]{2}-[0-9]{2,3}$",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/compare": {
      "get": {
        "tags": [
          "comparison"
        ],
        "summary": "Compare",
        "operationId": "compare_compare_get",
        "parameters": [
          {
            "name": "programIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Programids"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ComparisonScope",
              "default": "all"
            }
          },
          {
            "name": "semester",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssessmentType": {
        "type": "string",
        "enum": [
          "exam",
          "credit",
          "graded_credit",
          "coursework",
          "course_project",
          "state_exam"
        ],
        "title": "AssessmentType"
      },
      "CompareStatus": {
        "type": "string",
        "enum": [
          "both",
          "only_a",
          "only_b",
          "different"
        ],
        "title": "CompareStatus"
      },
      "ComparisonBlockResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "totalsA": {
            "$ref": "#/components/schemas/ComparisonTotalsResponse"
          },
          "totalsB": {
            "$ref": "#/components/schemas/ComparisonTotalsResponse"
          },
          "hoursDelta": {
            "type": "integer",
            "title": "Hoursdelta"
          },
          "creditsDelta": {
            "type": "string",
            "title": "Creditsdelta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "totalsA",
          "totalsB",
          "hoursDelta",
          "creditsDelta"
        ],
        "title": "ComparisonBlockResponse"
      },
      "ComparisonResponse": {
        "properties": {
          "programA": {
            "$ref": "#/components/schemas/ProgramSummaryResponse"
          },
          "programB": {
            "$ref": "#/components/schemas/ProgramSummaryResponse"
          },
          "scope": {
            "$ref": "#/components/schemas/ComparisonScope"
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRowResponse"
            },
            "type": "array",
            "title": "Rows"
          },
          "totalsA": {
            "$ref": "#/components/schemas/ComparisonTotalsResponse"
          },
          "totalsB": {
            "$ref": "#/components/schemas/ComparisonTotalsResponse"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/ComparisonBlockResponse"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "programA",
          "programB",
          "scope",
          "rows",
          "totalsA",
          "totalsB",
          "blocks"
        ],
        "title": "ComparisonResponse"
      },
      "ComparisonRowResponse": {
        "properties": {
          "discipline": {
            "$ref": "#/components/schemas/DisciplineResponse"
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "subjectGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectgroup"
          },
          "a": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkloadResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "b": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkloadResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CompareStatus"
          },
          "hoursDelta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hoursdelta"
          },
          "creditsDelta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creditsdelta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "discipline",
          "status"
        ],
        "title": "ComparisonRowResponse"
      },
      "ComparisonScope": {
        "type": "string",
        "enum": [
          "all",
          "semester"
        ],
        "title": "ComparisonScope"
      },
      "ComparisonTotalsResponse": {
        "properties": {
          "hours": {
            "type": "integer",
            "title": "Hours"
          },
          "credits": {
            "type": "string",
            "title": "Credits"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hours",
          "credits"
        ],
        "title": "ComparisonTotalsResponse"
      },
      "CurriculumItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "discipline": {
            "$ref": "#/components/schemas/DisciplineResponse"
          },
          "sourceName": {
            "type": "string",
            "title": "Sourcename"
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "hours": {
            "type": "integer",
            "title": "Hours"
          },
          "credits": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "assessmentTypes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssessmentType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessmenttypes"
          },
          "subjectGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectgroup"
          },
          "sourcePosition": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceposition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "discipline",
          "sourceName",
          "hours"
        ],
        "title": "CurriculumItemResponse"
      },
      "CurriculumResponse": {
        "properties": {
          "program": {
            "$ref": "#/components/schemas/ProgramSummaryResponse"
          },
          "curriculumId": {
            "type": "string",
            "title": "Curriculumid"
          },
          "educationYear": {
            "type": "integer",
            "title": "Educationyear"
          },
          "sourceUrl": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Sourceurl"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Capturedat"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CurriculumItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "program",
          "curriculumId",
          "educationYear",
          "sourceUrl",
          "capturedAt",
          "items"
        ],
        "title": "CurriculumResponse"
      },
      "DisciplineResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "normalizedName": {
            "type": "string",
            "title": "Normalizedname"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "normalizedName"
        ],
        "title": "DisciplineResponse"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "SOURCE_CONTRACT_ERROR",
          "CONTRACT_ERROR",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode"
      },
      "ErrorDetail": {
        "properties": {
          "path": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Path"
          },
          "message": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Message"
          },
          "type": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path",
          "message",
          "type"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "ProgramListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgramSummaryResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProgramListResponse"
      },
      "ProgramResponse": {
        "properties": {
          "program": {
            "$ref": "#/components/schemas/ProgramSummaryResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "program"
        ],
        "title": "ProgramResponse"
      },
      "ProgramSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "directionId": {
            "type": "string",
            "title": "Directionid"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "educationYear": {
            "type": "integer",
            "title": "Educationyear"
          },
          "studyPlanUrl": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Studyplanurl"
          },
          "sourceUrl": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Sourceurl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "directionId",
          "code",
          "name",
          "educationYear",
          "studyPlanUrl",
          "sourceUrl"
        ],
        "title": "ProgramSummaryResponse"
      },
      "WorkloadResponse": {
        "properties": {
          "semester": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "hours": {
            "type": "integer",
            "title": "Hours"
          },
          "credits": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "assessmentTypes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssessmentType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessmenttypes"
          },
          "subjectGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectgroup"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hours"
        ],
        "title": "WorkloadResponse"
      }
    }
  }
}